+++ packages/api-client/src/components/CodeInput/CodeInput.vue
@@ -274,18 +274,28 @@ export default {
 :deep(.cm-gutter + .cm-gutter .cm-gutterElement) {
   padding-left: 0 !important;
 }
+:deep(.cm-scroller) {
+  scrollbar-width: none;
+  -ms-overflow-style: none;
+  overflow: auto;
+}
+:deep(.cm-scroller::-webkit-scrollbar) {
+  width: 0;
+  height: 0;
+  display: none;
+}
 </style>
 <style>
 .cm-pill {
-  --tw-bg-base: var(--scalar-color-3);
+  --tw-bg-base: var(--scalar-color-1);
   color: var(--tw-bg-base);
-  padding: 1.5px 6px;
+  padding: 0 6px;
   margin: 0 6px;
   border-radius: 3px;
   display: inline-block;
   border-radius: 30px;
   font-size: var(--scalar-mini);
-  background: color-mix(in srgb, var(--tw-bg-base), transparent 90%) !important;
+  background: color-mix(in srgb, var(--tw-bg-base), transparent 94%) !important;
 }
 .dark-mode .cm-pill {
   background: color-mix(in srgb, var(--tw-bg-base), transparent 80%) !important;