+++ packages/api-client/src/views/Request/components/WorkspaceDropdown.vue
@@ -23,7 +23,8 @@ const createNewWorkspace = () =>
 </script>
 
 <template>
-  <div class="xl:min-h-header py-2.5 flex items-center border-b px-2.5 text-sm">
+  <div
+    class="xl:min-h-header py-2.5 flex items-center border-b-1/2 px-2.5 text-sm">
     <ScalarDropdown>
       <ScalarButton
         class="font-normal h-full justify-start py-1.5 px-1.5 text-c-1 hover:bg-b-2 w-fit"