+++ packages/api-reference-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @scalar/api-reference-react
 
+## 0.3.67
+
+### Patch Changes
+
+- Updated dependencies [9cc6311]
+  - @scalar/api-reference@1.24.69
+
 ## 0.3.66
 
 ### Patch Changes