+++ packages/play-button/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @scalar/play-button
 
+## 0.2.67
+
+### Patch Changes
+
+- Updated dependencies [d65dbcb]
+- Updated dependencies [9cc6311]
+  - @scalar/api-client@2.0.44
+  - @scalar/api-reference@1.24.69
+  - @scalar/oas-utils@0.2.25
+
 ## 0.2.66
 
 ### Patch Changes